move withRetry test to network pkg
